reorganize com crate for bindgen upgrade
