+++ crates/libs/com/src/Microsoft/mod.rs
@@ -1,0 +1,1 @@
+pub mod ServiceFabric;
